tezTracker MVP v2

+++ models/roll.go
@@ -2,8 +2,8 @@ package models
 
 type Roll struct {
 	Pkh        string `json:"pkh"`
-	Rolls      uint   `json:"rolls"`
+	Rolls      int64  `json:"rolls"`
 	BlockID    string `json:"block_id"`
 	Block      *Block `json:"block"` // This line is infered from column name "block_id".
-	BlockLevel uint   `json:"block_level"`
+	BlockLevel int64  `json:"block_level"`
 }
